Rev A
Fix bug, where data would come back as null and SQL nonnull, wouldn't allow us to insert it.

YahooFinance is really slow so and not consistent so there is another bug of missing data.
I set it where if there is no yahoo data to not update the table. The thing is that sometimes some values only come back as null not all dates thus causing a problem of missing time. A workout to this would be to change SQL DB to enable null and add the data either way.
I am okay with this at the moment as I just want to harvest data.

Perhaps a future implementation would be to check all tables of data and check for nulls in the data. if there is any null pull data from yahoo finance and update the table if it can find data.